simpleHttpExamplePython/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable
dapr==1.13.0
azure.monitor.opentelemetry==1.3.0
